packages/rig-pack-github/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rig-pack-github"
version = "0.1.0"
description = "RIG Pack: GitHub repository management"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
authors = [{name = "RIG Contributors"}]
dependencies = [
  "rig-core>=0.1.0",
  "PyGithub>=2.0.0"
]

[project.entry-points."rig.packs"]
rig-pack-github = "rig_pack_github.pack:PACK"

[tool.setuptools]
packages = ["rig_pack_github", "rig_pack_github.tools"]
